Perform username/email checks in a hook rather than custom function
Change-type: minor
